Refactor: 게임 진행 로직, 유효성 검사 로직 분리
